dietlens-backend/requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.9
Flask-SQLAlchemy==3.1.1
PyJWT==2.6.0
requests==2.28.2
Werkzeug==2.2.3
python-dotenv==1.0.0
Pillow==9.5.0
pymysql==1.0.3
cryptography==40.0.2
Requests
flask_sqlalchemy
pyjwt
pymysql
flask-cors==4.0.0